+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.11.3
 
 - Key changes:
-  - Fixed deduplication for negative non-regexp filters (previously, those would be let through without request modifications);
+  - Fixed deduplication for negative non-regexp filters (previously, some of those would be let through without request modifications);
   - Internal refactoring:
     - Moved acl + lf logic to a new internal package `querymodifier`;
   - Added more tests.